// src/app/shared/client/referencia-vehiculo.ts
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 * created by oscarcisneros oscarcisneros8@gmail.com
 */
import {ReferenceDTO} from './referenceDTO';
export class ReferenciaVehiculo extends ReferenceDTO {
  correlativoReferencia: number;
  guardada: boolean;
  empresaFinancia: string;
  marca: string;
  modelo: string;
  placaVehiculo: string;
  placa: string;
  saldoFinanciamiento: number;
}